pyproject.toml
[project]
name = "hardcopy"
version = "0.0.0"
description = "Copy tool with validation for Windows, wrapping Robocopy."
authors = [
    {name = "bjhardcastle", email = "ben.hardcastle@alleninstitute.org"},
]
dependencies = [
    "typer[all]>=0.7.0",
    "crc32c>=2.0",
    "backports-strenum>=1.0.3",
]
requires-python = ">=3.7"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]

[tool.pdm.scripts]
# usage: pdm run <script> [args]
isort = "isort src"
blue = "blue src"
pytest = "pytest --cov"
prebuild = {composite = ["isort", "blue", "pytest"]}
bump = "bump pyproject.toml" # defaults to minor; `pdm run bump -M` to bump major
build = {composite = ["prebuild", "bump", "pdm build"]}
dry-run = {composite = ["prebuild", "pdm build", "pdm publish --no-build --repository https://test.pypi.org/simple"]}
publish = {composite = ["prebuild", "pdm build", "pdm publish --no-build"]}

[tool.pdm.source]
# name = "hardcopy"

[tool.pdm.dev-dependencies]
dev = [
    "blue>=0.9.1",
    "pytest>=7.2.2",
    "isort>=5.11.5",
    "mypy>=1.1.1",
    "coverage[toml]>=7.2.2",
    "pdm>=2.4.9",
    "pytest-cov>=4.0.0",
    "bump>=1.3.2",
    "types-backports>=0.1.3",
]

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
    "--doctest-modules",
    # "--pdb",
    # "--pdbcls=IPython.terminal.debugger:TerminalPdb",
]

doctest_optionflags = [
    "NORMALIZE_WHITESPACE",
    "IGNORE_EXCEPTION_DETAIL",
    "ELLIPSIS",
]

testpaths = [
    # "tests",
    "src",
]

[tool.coverage.paths]
source = ["src"]

[tool.coverage.run]
branch = true
source = ["hardcopy"]

[tool.coverage.report]
show_missing = true